pcxt: implement SS save/restore for per-process kernel stacks
ISR (switch.S) and syscall (trap.S) handlers now save GP registers on
the interrupted stack (user or kernel), then switch SS:SP to the
per-process kernel stack (SS=0).  Only the interrupted SS:SP pair
(4 bytes) lives on the kernel stack; the 24-byte GP+IRET frame stays
on the interrupted stack.  On exit, user SS:SP is restored and IRET
pops FLAGS/CS/IP from there.  If already on kernel stack (SS=0, e.g.
PID 0 idle loop), the stack switch is skipped.

elf16_loader now builds the initial frame in two parts: the 24-byte
GP+IRET frame on the user stack (SS=proc_seg, segment-relative SP)
and the 4-byte user_SS:SP on the kernel stack.  pcb->sp points to
the kernel stack.  arch_build_initial_frame (PID 0 only) builds both
parts on the same kernel stack for the SS=0 case.

Each process slot gets kernel_stack_top assigned in proc_init()
(0xF000 + (slot+1)*1024).  i16_current_ksp global is updated on
context switch so the ISR can load the kernel SP atomically via
the 8086 SS-load interrupt inhibit.

Also fix page_id_linear() and linear_page_id() return/parameter
types from uintptr_t (16-bit on i16) to uint32_t, fixing page pool
address truncation for pages above 64 KB.

Verified: builds cleanly, boots on QEMU to scheduler with correct
page pool address (0x21000-0x9EFFF, 126 pages).

+++ src/kernel/mm/page.c
@@ -69,7 +69,7 @@ static inline uint32_t page_id_linear(page_id_t id) {
   return (uint32_t)id * PAGE_SIZE;
 }
 
-static inline page_id_t linear_page_id(uintptr_t addr) {
+static inline page_id_t linear_page_id(uint32_t addr) {
   return (page_id_t)(addr / PAGE_SIZE);
 }
